Integrted new layout for multiple users
