+++ public/js/wp-agora-io-public.js
@@ -28,7 +28,84 @@
 })(jQuery,'smartresize');
 
 
+
+function changeStreamSource (deviceIndex, deviceType) {
+  AgoraRTC.Logger.info('Switching stream sources for: ' + deviceType);
+  var deviceId;
+  var existingStream = false;
+  
+  if (deviceType === "video") {
+    deviceId = window.devices.cameras[deviceIndex].deviceId
+  }
+
+  if(deviceType === "audio") {
+    deviceId = window.devices.mics[deviceIndex].deviceId;
+  }
+
+  window.localStreams.camera.stream.switchDevice(deviceType, deviceId, function(){
+    AgoraRTC.Logger.info('successfully switched to new device with id: ' + JSON.stringify(deviceId));
+    // set the active device ids
+    if(deviceType === "audio") {
+      window.localStreams.camera.micId = deviceId;
+    } else if (deviceType === "video") {
+      window.localStreams.camera.camId = deviceId;
+    } else {
+      AgoraRTC.Logger.warning("unable to determine deviceType: " + deviceType);
+    }
+  }, function(){
+    AgoraRTC.Logger.error('failed to switch to new device with id: ' + JSON.stringify(deviceId));
+  });
+}
+
+// helper methods
+function getCameraDevices() {
+  AgoraRTC.Logger.info("Checking for Camera window.devices.....")
+  window.agoraClient.getCameras (function(cameras) {
+    window.devices.cameras = cameras; // store cameras array
+    cameras.forEach(function(camera, i){
+      var name = camera.label.split('(')[0];
+      var optionId = 'camera_' + i;
+      var deviceId = camera.deviceId;
+      if(i === 0 && window.localStreams.camera.camId === ''){
+        window.localStreams.camera.camId = deviceId;
+      }
+      jQuery('#camera-list').append('<a class="dropdown-item" id="' + optionId + '">' + name + '</a>');
+    });
+    jQuery('#camera-list a').click(function(event) {
+      var index = event.target.id.split('_')[1];
+      changeStreamSource (index, "video");
+    });
+  });
+}
+
+function getMicDevices() {
+  AgoraRTC.Logger.info("Checking for Mic window.devices.....")
+  window.agoraClient.getRecordingDevices(function(mics) {
+    window.devices.mics = mics; // store mics array
+    mics.forEach(function(mic, i){
+      var name = mic.label.split('(')[0];
+      var optionId = 'mic_' + i;
+      var deviceId = mic.deviceId;
+      if(i === 0 && window.localStreams.camera.micId === ''){
+        window.localStreams.camera.micId = deviceId;
+      }
+      if(name.split('Default - ')[1] != undefined) {
+        name = '[Default Device]' // rename the default mic - only appears on Chrome & Opera
+      }
+      jQuery('#mic-list').append('<a class="dropdown-item" id="' + optionId + '">' + name + '</a>');
+    }); 
+    jQuery('#mic-list a').click(function(event) {
+      var index = event.target.id.split('_')[1];
+      changeStreamSource (index, "audio");
+    });
+  });
+}
+
+
 window.AGORA_UTILS = {
+
+  getMicDevices: getMicDevices,
+  getCameraDevices: getCameraDevices,
 
   agoraApiRequest: function (endpoint_url, endpoint_data) {
     var ajaxRequestParams = {
