+++ public/views/agora-fullscreen-broadcast.php
@@ -29,75 +29,76 @@ $current_user       = wp_get_current_user();
         </div>
       </div>
 
-      <?php require_once "parts/footer.php" ?>
+      <?php require_once "parts/footer-broadcast.php" ?>
     </section>
     
-  </div>
 
-  <!-- RTMP Config Modal -->
-  <div class="modal fade slideInLeft animated" id="addRtmpConfigModal" tabindex="-1" role="dialog" aria-labelledby="rtmpConfigLabel" aria-hidden="true" data-keyboard=true>
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header">
-          <h5 class="modal-title" id="rtmpConfigLabel"><i class="fas fa-sliders-h"></i></h5>
-          <button type="button" class="close" data-dismiss="modal" data-reset="modal" aria-label="Close">
-            <span aria-hidden="true">&times;</span>
-          </button>
-        </div>
-        <div class="modal-body">
-          <form id="rtmp-config" action="" method="post" onSubmit="return false;">
-            <div class="form-group">
-              <label for="input_rtmp_url">RTMP Server URL</label>
-              <input type="url" class="form-control" id="input_rtmp_url" placeholder="Enter the RTMP Server URL" value="" required />
-            </div>
-            <div class="form-group">
-              <label for="input_private_key">Stream key</label>
-              <input type="text" class="form-control" id="input_private_key" placeholder="Enter stream key" required />
-            </div>
-            <input type="submit" value="Start RTMP" style="position:fixed; top:-999999px">
-          </form>
-        </div>
-        <div class="modal-footer">
-          <span id="rtmp-error-msg" class="error text-danger" style="display: none">Please complete the information!</span>
-          <button type="button" id="start-RTMP-broadcast" class="btn btn-primary">
-            <i class="fas fa-satellite-dish"></i>
-          </button>
+    <!-- RTMP Config Modal -->
+    <div class="modal fade slideInLeft animated" id="addRtmpConfigModal" tabindex="-1" role="dialog" aria-labelledby="rtmpConfigLabel" aria-hidden="true" data-keyboard=true>
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="rtmpConfigLabel"><i class="fas fa-sliders-h"></i></h5>
+            <button type="button" class="close" data-dismiss="modal" data-reset="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="modal-body">
+            <form id="rtmp-config" action="" method="post" onSubmit="return false;">
+              <div class="form-group">
+                <label for="input_rtmp_url">RTMP Server URL</label>
+                <input type="url" class="form-control" id="input_rtmp_url" placeholder="Enter the RTMP Server URL" value="" required />
+              </div>
+              <div class="form-group">
+                <label for="input_private_key">Stream key</label>
+                <input type="text" class="form-control" id="input_private_key" placeholder="Enter stream key" required />
+              </div>
+              <input type="submit" value="Start RTMP" style="position:fixed; top:-999999px">
+            </form>
+          </div>
+          <div class="modal-footer">
+            <span id="rtmp-error-msg" class="error text-danger" style="display: none">Please complete the information!</span>
+            <button type="button" id="start-RTMP-broadcast" class="btn btn-primary">
+              <i class="fas fa-satellite-dish"></i>
+            </button>
+          </div>
         </div>
       </div>
     </div>
-  </div>
-  <!-- end Modal -->
+    <!-- end Modal -->
 
-  <!-- External Injest Url Modal -->
-  <div class="modal fade slideInLeft animated" id="add-external-source-modal" tabindex="-1" role="dialog" aria-labelledby="add-external-source-url-label" aria-hidden="true" data-keyboard=true>
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header">
-          <h5 class="modal-title" id="add-external-source-url-label">
-            <i class="fas fa-broadcast-tower"></i> [add external url]
-          </h5>
-          <button id="hide-external-url-modal" type="button" class="close" data-dismiss="modal" data-reset="modal" aria-label="Close">
-            <span aria-hidden="true">&times;</span>
-          </button>
-        </div>
-        <div class="modal-body">
-          <form id="external-inject-config">
-            <div class="form-group">
-              <label for="input_external_url">External URL</label>
-              <input type="url" class="form-control" id="input_external_url" placeholder="Enter the external URL" required>
-            </div>
-          </form>
-        </div>
-        <div class="modal-footer">
-          <span id="external-url-error" class="error text-danger" style="display: none">Please enter a valid external URL</span>
-          <button type="button" id="add-external-stream" class="btn btn-primary">
-              <i id="add-rtmp-icon" class="fas fa-plug"></i>  
-          </button>
+    <!-- External Injest Url Modal -->
+    <div class="modal fade slideInLeft animated" id="add-external-source-modal" tabindex="-1" role="dialog" aria-labelledby="add-external-source-url-label" aria-hidden="true" data-keyboard=true>
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="add-external-source-url-label">
+              <i class="fas fa-broadcast-tower"></i> [add external url]
+            </h5>
+            <button id="hide-external-url-modal" type="button" class="close" data-dismiss="modal" data-reset="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="modal-body">
+            <form id="external-inject-config">
+              <div class="form-group">
+                <label for="input_external_url">External URL</label>
+                <input type="url" class="form-control" id="input_external_url" placeholder="Enter the external URL" required>
+              </div>
+            </form>
+          </div>
+          <div class="modal-footer">
+            <span id="external-url-error" class="error text-danger" style="display: none">Please enter a valid external URL</span>
+            <button type="button" id="add-external-stream" class="btn btn-primary">
+                <i id="add-rtmp-icon" class="fas fa-plug"></i>  
+            </button>
+          </div>
         </div>
       </div>
     </div>
+    <!-- end Modal -->
+    
   </div>
-  <!-- end Modal -->
 
   <?php wp_footer(); ?>
   <script>
@@ -127,25 +128,6 @@ $current_user       = wp_get_current_user();
       window.cameraVideoProfile = '<?php echo $instance['videoprofile'] ?>';
       window.screenVideoProfile = '<?php echo $instance['screenprofile'] ?>';
 
-      // keep track of streams
-      window.localStreams = {
-        uid: '',
-        camera: {
-          camId: '',
-          micId: '',
-          stream: {}
-        },
-        screen: {
-          id: "",
-          stream: {}
-        }
-      };
-
-      // keep track of devices
-      window.devices = {
-        cameras: [],
-        mics: []
-      }
 
       window.externalBroadcastUrl = '';
       // default config for rtmp
