--- public/views/parts/footer.php
@@ -1,7 +1,0 @@
-<footer class="agora-footer">
-	<div class="buttons-botton">
-		<i class="btnIcon icon-mic"></i>
-		<i class="btnIcon icon-users"></i>
-		<i class="btnIcon icon-feed"></i>
-	</div>
-</footer>
