+++ public/views/parts/header-controls.php
@@ -1,12 +1,12 @@
 <div class="buttons-top">
   <div class="left-button">
-    <button class= "wordpress">
-      Wordpress
-    </button>
+    <div class= "room-title">
+      <?php wp_title(); ?> | <span id="count-users">1</span> <i id="count-users-icon" class="fas fa-users"></i>
+    </div>
   </div>
   <div class="right-buttons">
 
-    <button class= "leave-meeting">
+    <button id="exit-btn" class= "leave-meeting">
       Leave Meeting
     </button>
     <i class="icon-cog"></i>
